Added explicit types and key type
I thought it was a bit ambiguous on what 'type' of key was generated in C#, was it a `string`? was it an '`nt`? was it a `float`?  -- from the frontpage, we can only see `var` which is not helpful, or the key value, which is still a bit ambiguous. 

So the reader must explore the code, and explore the text at the bottom.

This is why I think its helpful to add explicit types. I also added a new FAQ answer to further make this clear to the reader.